Triggers the menu when pressing the enter key
